requirements.txt
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
python-telegram-bot==21.1.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
